double to long long

+++ 0x00-hello_world/6-size.c
@@ -7,13 +7,9 @@
 int main(void)
 {
 char a;
-
 int b;
-
 long int c;
-
-double d;
-
+long long d;
 float e;
 
 printf("Size of a char: %lu byte(s)\n", sizeof(a));
